requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
copilotkit==0.1.39
crashtest==0.4.1
cryptography==43.0.3
distlib==0.3.9
distro==1.9.0
dulwich==0.22.7
ecdsa==0.19.1
fastapi==0.115.10
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.20.0
keyring==25.6.0
langchain==0.3.19
langchain-anthropic==0.3.8
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langgraph==0.3.2
langgraph-api==0.0.27
langgraph-checkpoint==2.0.16
langgraph-cli==0.1.74
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langsmith==0.3.11
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.32.0
numpy==1.26.4
openai==1.65.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
partialjson==0.0.8
passlib==1.7.4
pbs-installer==2025.2.12
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.1.1
poetry-core==2.1.1
propcache==0.3.0
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.4.8
pycparser==2.22
pydantic~=2.7
pydantic_core==2.27.2
pydeck==0.9.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sse-starlette==2.1.3
starlette==0.46.0
streamlit==1.44.0
structlog==23.3.0
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
trove-classifiers==2025.2.18.16
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
watchdog==6.0.0
watchfiles==1.0.4
yarl==1.18.3
zstandard==0.23.0
python-multipart